My solution to Free Code Camp's Caesar's Cipher
This one was kinda tough, so I did the honorable thing and make it even tougher on myself by using Regex!! I really am trying to learn Regex and when I see something I can use it on, I jump at it. Let me walk you thru my solution:

Found it easier to save the Regex as variables, I highly recommend using http://regexr.com/ to test out your Regex.

I decided to make a for loop, don't forget since you are looping thru a string (as opposed to an array) your sting length must be ".length -1"

 The first If statement I'm looking for spaces and punctuation, this can't be in the other Regex since instead of adding or subtracting 13 I'm simply passing them on to the final string.
Now since you are only using the letters from A to Z you simply can't add 13 to each letter. You only add 13 to the first 13 letters in the Alphabet, the last 13 you subtract 13. The ".test" is a regex test function I used here.  Also notice the code in the "if" statement. You should know by now the "if" statement tests for a true statement, there is no need for you to add "== true" to the code. If the regex tests true it outputs a Boolean true. If it test true I then grab the letter, convert is to ASCII , add 13, then convert it back to a letter and add it to the "newstr" string. Since you already checked for spaces and punctuation, plus the letters from A-M, the else statement doesn't need any logic, the only think left in the test is letters after M which you then subtract 13 from. Also a note about the "String.fromCharCode()" function. It expects comma separated numbers it can convert to a string. It will not work with variables! So I simply used it to convert each digit.

Yes there were easier ways to do this, however I learned alot doing this the "hard" way.